refactor(common): promote 'codec' module to  directory
