+++ mithril-common/src/crypto_helper/codec/json_hex.rs
@@ -5,11 +5,6 @@ use serde::de::DeserializeOwned;
 use serde::Serialize;
 use thiserror::Error;
 
-use anyhow::anyhow;
-use bech32::{self, Bech32, Hrp};
-
-use crate::StdResult;
-
 /// Error raised when the encoding or decoding fails
 #[derive(Error, Debug)]
 #[error("Codec error: {msg}")]
@@ -61,18 +56,11 @@ where
     })
 }
 
-/// Encode to bech32 given Human Readable Part (hrp) and data
-pub fn encode_bech32(human_readable_part: &str, data: &[u8]) -> StdResult<String> {
-    let human_readable_part = Hrp::parse(human_readable_part).map_err(|e| anyhow!(e))?;
-    bech32::encode::<Bech32>(human_readable_part, data).map_err(|e| anyhow!(e))
-}
-
 #[cfg(test)]
 mod tests {
-    use hex::FromHex;
     use serde::{Deserialize, Serialize};
 
-    use super::{encode_bech32, key_decode_hex, key_encode_hex};
+    use super::{key_decode_hex, key_encode_hex};
 
     #[derive(Debug, PartialEq, Serialize, Deserialize)]
     struct TestSerialize {
@@ -101,17 +89,5 @@ mod tests {
         let test_to_serialize_restored = key_decode_hex(&format!(" \r{test_to_serialize_hex} \n"))
             .expect("unexpected hex decoding error");
         assert_eq!(test_to_serialize, test_to_serialize_restored);
-    }
-
-    #[test]
-    fn test_bech32_encode() {
-        let hrp = "pool";
-        let data =
-            Vec::from_hex("edfa208d441511f9595ba80e8f3a7b07b6a80cbc9dda9d8e9d1dc039").unwrap();
-        let encoded_data = encode_bech32(hrp, &data).unwrap();
-        let expected_encoded_data =
-            "pool1ahazpr2yz5gljk2m4q8g7wnmq7m2sr9unhdfmr5arhqrjnntwdz".to_string();
-
-        assert_eq!(expected_encoded_data, encoded_data);
     }
 }
